test(js): use nodejs testing instead of ava
