--- iroh-js/__test__/_force-exit.mjs
@@ -1,6 +1,0 @@
-import process from 'node:process';
-import { registerCompletionHandler } from 'ava';
-
-registerCompletionHandler(() => {
-  process.exit();
-});
