--- iroh-js/__test__/net.spec.mjs
@@ -1,10 +1,0 @@
-import test from 'ava'
-
-import { Iroh } from '../index.js'
-
-test('node status', async (t) => {
-  const iroh = await Iroh.memory()
-
-  const nodeId = await iroh.net.nodeId()
-  t.truthy(nodeId)
-})
